server/requirements.txt
flask==2.3.2
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1
sqlalchemy==1.3.24
sqlalchemy-serializer==1.4.1
flask-cors==3.0.10
flask-restful
werkzeug==2.3.3
Flask-Session==0.4
flask-restful==0.3.9
python-dotenv==1.0.0

